requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
filelock==3.8.0
Markdown==3.4.1
packaging==21.3
phonenumbers==8.13.0
Pillow==9.3.0
platformdirs==2.5.4
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tox==3.27.0
# virtualenv==20.16.7
